Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -16,7 +16,7 @@ int main(){
         printf("%d", mult);
     }
     else if (c=='/'){
-        int div=a/b;
+        float div=a/b;
         printf("%.1f", div);
     }
     else {
